Add files via upload
Completed an exciting project on data analysis using Microsoft Excel. The goal was to extract the valuable insights from the data and get better understanding of complaints and how to effectively resolve the issues. In this project I use Data Analysis, Data Cleaning & Data Visualization process to get this result.

Problems that this project solves:-
1.     Do consumer complaints show any seasonal patterns?
2.     Which products present the most complaints?
3. What are its most common issues?
4.     How are complaints typically resolved?
5. Which state has most complaints?
6. Complaints raised per month?